requirements.txt
# 核心框架
langgraph>=0.2.0
openai>=1.60.0
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.10.0

# 图片处理
Pillow>=11.0.0

# CLI 接口
click>=8.2.0
rich>=14.0.0
